+++ CHANGES.rst
@@ -1,7 +1,9 @@
 v0.4
 ====
 - Don't use consants FLAT, CYLINDRICAL, SPHERICAL. Instead use 'f', 'c', 's'
-- Don't use iterative 0, 1, 2, 3. Instead use: 'false', 'gmres', 'bicgstab', 'tfqmr'
+- For sundials backend: Don't use iterative 0, 1, 2, 3. Instead use:
+   - linear_solver: {'default', 'dense', 'banded', 'gmres', 'bicgstab', 'tfqmr'}
+   - iter_type: {'functional', 'newton'}
 - Drop constants GEOM_ORDER, DENSE, BANDED, SPARSE, GEOMS